requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-common==1.1.27
azure-core==1.13.0
azure-functions==1.7.0
azure-servicebus==7.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msal==1.11.0
msrest==0.6.21
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.1.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
uamqp==1.4.0
urllib3==1.26.4